==> Phase_1/notebooks/1. Initial_Merging.ipynb <==
import pandas as pd
from pathlib import Path


# Paths
file1 = Path('C:/Users/HP/OneDrive/Desktop/VERO_code/Phase_1/data/raw')
file2 = Path('C:/Users/HP/OneDrive/Desktop/VERO_code/Phase_1/data/raw/Final Dataset')

# Load both
df1 = pd.read_excel(file1 / "codige.xlsx")
df2 = pd.read_excel(file2 / "codige_final.xlsx")



# Merge (stack) them by rows
merged_raw = pd.concat([df1, df2], ignore_index=True, sort=False)

# Save to a new file
merged_raw.to_excel("C:/Users/HP/OneDrive/Desktop/VERO_code/Phase_1/data/raw/merged_codige.xlsx", index=False)

print("Merging complete. File saved as merged_codige.xlsx")


# Paths
file1 = Path('C:/Users/HP/OneDrive/Desktop/VERO_code/Phase_1/data/raw')
file2 = Path('C:/Users/HP/OneDrive/Desktop/VERO_code/Phase_1/data/raw/Final Dataset')

# Load both
df1 = pd.read_excel(file1 / "codige_trattamento_oncologico.xlsx")
df2 = pd.read_excel(file2 / "codige_trattamento_oncologico.xlsx")



# Merge (stack) them by rows
merged_raw = pd.concat([df1, df2], ignore_index=True, sort=False)

# Save to a new file
merged_raw.to_excel("C:/Users/HP/OneDrive/Desktop/VERO_code/Phase_1/data/raw/merged_codige_trattamento_oncologico.xlsx", index=False)

print("Merging complete. File saved as merged_codige_trattamento_oncologico.xlsx")


# Paths
file1 = Path('C:/Users/HP/OneDrive/Desktop/VERO_code/Phase_1/data/raw')
file2 = Path('C:/Users/HP/OneDrive/Desktop/VERO_code/Phase_1/data/raw/Final Dataset')

# Load both
df1 = pd.read_excel(file1 / "codige_ricoveri.xlsx")
df2 = pd.read_excel(file2 / "codige_ricoveri.xlsx")



# Merge (stack) them by rows
merged_raw = pd.concat([df1, df2], ignore_index=True, sort=False)

# Save to a new file
merged_raw.to_excel("C:/Users/HP/OneDrive/Desktop/VERO_code/Phase_1/data/raw/merged_codige_ricoveri.xlsx", index=False)

print("Merging complete. File saved as merged_codige_ricoveri.xlsx")


# Paths
file1 = Path('C:/Users/HP/OneDrive/Desktop/VERO_code/Phase_1/data/raw')
file2 = Path('C:/Users/HP/OneDrive/Desktop/VERO_code/Phase_1/data/raw/Final Dataset')

# Load both
df1 = pd.read_excel(file1 / "codige_comorbidità.xlsx")
df2 = pd.read_excel(file2 / "codige_comorbilità.xlsx")



# Merge (stack) them by rows
merged_raw = pd.concat([df1, df2], ignore_index=True, sort=False)

# Save to a new file
merged_raw.to_excel("C:/Users/HP/OneDrive/Desktop/VERO_code/Phase_1/data/raw/merged_codige_comorbidità.xlsx", index=False)

print("Merging complete. File saved as merged_codige_comorbidità.xlsx")


# Paths
file1 = Path('C:/Users/HP/OneDrive/Desktop/VERO_code/Phase_1/data/raw')
file2 = Path('C:/Users/HP/OneDrive/Desktop/VERO_code/Phase_1/data/raw/Final Dataset')

# Load both
df1 = pd.read_excel(file1 / "codige_adr_clean.xlsx")
df2 = pd.read_excel(file2 / "codige_adr_clean (1).xlsx")



# Merge (stack) them by rows
merged_raw = pd.concat([df1, df2], ignore_index=True, sort=False)

# Save to a new file
merged_raw.to_excel("C:/Users/HP/OneDrive/Desktop/VERO_code/Phase_1/data/raw/merged_codige_adr_clean.xlsx", index=False)

print("Merging complete. File saved as merged_codige_adr_clean.xlsx")
